Refactored CDK classes to be shared between components and stacks.
