+++ pipeline/aws_infra/mypy.ini
@@ -9,3 +9,4 @@ strict_equality = True
 enable_error_code = explicit-override,truthy-bool,truthy-iterable,possibly-undefined
 follow_imports = skip
 exclude = venv
+mypy_path = ../../shared_aws_infra/
