+++ shared_aws_infra/.gitignore
@@ -1,0 +1,115 @@
+# CDK files
+## asset staging directory
+.cdk.staging
+cdk.out
+
+## CLI installation artifacts
+node_modules/
+package.json
+package-lock.json
+
+# Python files
+## Byte-compiled / optimized / DLL files
+__pycache__/
+*.py[cod]
+*$py.class
+
+## C extensions
+*.so
+
+## Distribution / packaging
+.Python
+build/
+develop-eggs/
+dist/
+downloads/
+eggs/
+.eggs/
+lib/
+lib64/
+parts/
+sdist/
+var/
+wheels/
+share/python-wheels/
+*.egg-info/
+.installed.cfg
+*.egg
+MANIFEST
+
+## PyInstaller
+#  Usually these files are written by a python script from a template
+#  before PyInstaller builds the exe, so as to inject date/other infos into it.
+*.manifest
+*.spec
+
+## package installer logs
+pip-log.txt
+pip-delete-this-directory.txt
+
+## Unit test / coverage reports
+htmlcov/
+.tox/
+.nox/
+.coverage
+.coverage.*
+.cache
+nosetests.xml
+coverage.xml
+*.cover
+*.py,cover
+.hypothesis/
+.pytest_cache/
+cover/
+
+## Translations
+*.mo
+*.pot
+
+## Sphinx documentation
+docs/_build/
+
+## PyBuilder
+.pybuilder/
+target/
+
+## IPython
+profile_default/
+ipython_config.py
+
+## Celery stuff
+celerybeat-schedule
+celerybeat.pid
+
+## (virtual) environments
+.env
+.venv
+env/
+venv/
+ENV/
+env.bak/
+venv.bak/
+
+## Spyder project settings
+.spyderproject
+.spyproject
+
+## Rope project settings
+.ropeproject
+
+## mkdocs documentation
+/site
+
+## mypy
+.mypy_cache/
+.dmypy.json
+dmypy.json
+
+## Pyre type checker
+.pyre/
+
+## pytype static type analyzer
+.pytype/
+
+## Cython debug symbols
+cython_debug/
