+++ shared_aws_infra/Makefile
@@ -1,0 +1,25 @@
+.PHONY: check-venv-active check-node
+SUPPORTED_NODE := ^v18\.
+
+check-venv-active:
+ifeq ($(VIRTUAL_ENV),)
+	@echo 'No active python virtual environment found.'\
+		  'Please active the virtual environment first by running `source venv/bin/activate`,'\
+		  'or read README.md for instructions on how to set up a new one.'
+	@exit 1
+else
+	@:
+endif
+
+check-node:
+	@$(eval NODE_VERSION=$(shell node --version))
+ifeq ($(shell node --version | grep -P "${SUPPORTED_NODE}"),)
+	@echo 'Node version "${NODE_VERSION}" not supported!'\
+		  'Change the active node version (use nvm) to match "${SUPPORTED_NODE}".'
+	@exit 1
+else
+	@:
+endif
+
+python-dependencies:
+	pip install -r requirements.txt
