+++ shared_aws_infra/shared_cdk_classes/pavi_ecr_repo.py
